src/chatbot/studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
sqlalchemy
pydantic
typing
python-dotenv
baml-py
